+++ thoralf-plugin.cabal
@@ -62,7 +62,7 @@ Test-suite tests
   default-language:      Haskell2010
   hs-source-dirs:        test
   main-is:               Main.hs
-  other-modules:
+  other-modules:         Nat, FiniteMaps, RowTypes
   ghc-options:           -Wall -fplugin ThoralfPlugin.Plugin
   build-depends:         base
                        , thoralf-plugin
